Migrate SomePointer, isGlobalPointer to more specific modules
